+++ go/2019/day07/part1.go
@@ -6,9 +6,9 @@ import (
 	"gonum.org/v1/gonum/stat/combin"
 )
 
-func FindMaxOutput(originalProgram *lib.Program) int {
+func FindMaxOutput(originalProgram *lib.Program) int64 {
 	permGen := combin.NewPermutationGenerator(5, 5)
-	max := 0
+	max := int64(0)
 
 	for permGen.Next() {
 		phases := permGen.Permutation(nil)
@@ -23,17 +23,17 @@ func FindMaxOutput(originalProgram *lib.Program) int {
 	return max
 }
 
-func GetSignal(originalProgram *lib.Program, phases []int) int {
-	previousInput := 0
+func GetSignal(originalProgram *lib.Program, phases []int) int64 {
+	previousInput := int64(0)
 
 	for i := 0; i < 5; i++ {
-		programMemory := make([]int, len(originalProgram.Data))
+		programMemory := make([]int64, len(originalProgram.Data))
 		copy(programMemory, originalProgram.Data)
-		program := lib.NewProgram(fmt.Sprintf("Program %d", i), programMemory, make(chan int), make(chan int))
+		program := lib.NewProgram(fmt.Sprintf("Program %d", i), programMemory, make(chan int64), make(chan int64))
 
 		go program.Run()
 
-		program.Input <- phases[i]
+		program.Input <- int64(phases[i])
 		program.Input <- previousInput
 		previousInput = <-program.Output
 	}
